+++ src/main/java/net/sodiumstudio/dwmg/entities/hmag/EntityBefriendedWitherSkeletonGirl.java
@@ -85,14 +85,14 @@ public class EntityBefriendedWitherSkeletonGirl extends WitherSkeletonGirlEntity
 
 	protected static final EntityDataAccessor<Optional<UUID>> DATA_OWNERUUID = SynchedEntityData
 			.defineId(EntityBefriendedWitherSkeletonGirl.class, EntityDataSerializers.OPTIONAL_UUID);
-	protected static final EntityDataAccessor<Byte> DATA_AISTATE = SynchedEntityData
-			.defineId(EntityBefriendedWitherSkeletonGirl.class, EntityDataSerializers.BYTE);
+	protected static final EntityDataAccessor<Integer> DATA_AISTATE = SynchedEntityData
+			.defineId(EntityBefriendedWitherSkeletonGirl.class, EntityDataSerializers.INT);
 
 	@Override
 	protected void defineSynchedData() {
 		super.defineSynchedData();
 		entityData.define(DATA_OWNERUUID, Optional.empty());
-		entityData.define(DATA_AISTATE, (byte) 0);
+		entityData.define(DATA_AISTATE, 0);
 	}
 
 	@Override
@@ -101,7 +101,7 @@ public class EntityBefriendedWitherSkeletonGirl extends WitherSkeletonGirlEntity
 	}
 
 	@Override
-	public EntityDataAccessor<Byte> getAIStateData() {
+	public EntityDataAccessor<Integer> getAIStateData() {
 		return DATA_AISTATE;
 	}
